llm-backend/test/rag-gemini-llama-index/requirements.txt
llama-index-vector-stores-qdrant==0.3.0
llama-index-llms-ollama==0.3.1
llama-index-embeddings-fastembed==0.2.0
llama-index-embeddings-ollama==0.3.0
llama-index-postprocessor-rankgpt-rerank == 0.2.0